disable manual pan, zoom, rotate on possum, move out of subdirectory
